+++ next.config.js
@@ -6,14 +6,6 @@ const nextConfig = {
     locales: ["en"],
     defaultLocale: "en",
   },
-  async rewrites() {
-    return [
-      {
-        source: "/proxy/:path*",
-        destination: "https://p.scdn.co/mp3-preview/:path*",
-      },
-    ];
-  },
 };
 
 module.exports = nextConfig;