+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.0] - Dec 26, 2023
+## [2.1.0] - Dec 27, 2023
+
+### Added
+
+- Add a new audio proxy api route
+
+### Changed
+
+- Use an audio proxy api route instead of Next.js rewrites
+
+### Fixed
+
+- Fix the header logo src not being relative
+
+## [2.0.0] - Dec 27, 2023
 
 ### Added